improve retry handling
